Earth 45 refactor deploy script (#45)
* earth-45: Resolves PSScriptAnalyzer issues.

* earth-45: Fixes more PowerShell warnings and adds a build.ps1 script for checking things like PowerShell lint.

* earth-45: Fix build.ps1 invocation.

* earth-45: Testing a lint failure.

* earth-45: Fix lint failure.

* earth-45: Removes the verbose copy of git hooks.

+++ azure/subscription-deprovision.ps1
@@ -1,4 +1,4 @@
-[CmdletBinding()]
+﻿[CmdletBinding()]
 param (
     [Parameter(Mandatory=$true)]
     [String]
@@ -16,19 +16,26 @@ $SubscriptionId = (az account show --subscription $AzureSubscriptionName | Conve
 Write-Information "Deprovisioning subscription $AzureSubscriptionName (id: $SubscriptionId) ..."
 
 function Remove-DeployerServicePrincipal {
-    $SubscriptionDeploymentServicePricipalName = "$AzureSubscriptionName-earth-deployer".ToLower()
-    
-    $ServicePrincipalId = (
-        az ad sp list `
-            --filter "displayname eq '$SubscriptionDeploymentServicePricipalName'" | ConvertFrom-Json
-    ).appId
+    [CmdletBinding(SupportsShouldProcess)]
+    Param()
 
-    if ($ServicePrincipalId) {
-        Write-Information "Deprovisioning $SubscriptionDeploymentServicePricipalName Service Principal (id: $ServicePrincipalId) ..."
+    process {
+        if($PSCmdlet.ShouldProcess($AzureSubscriptionName)) {
+            $SubscriptionDeploymentServicePricipalName = "$AzureSubscriptionName-earth-deployer".ToLower()
 
-        az ad sp delete --id $ServicePrincipalId
-        
-        Write-Information "Service Principal $SubscriptionDeploymentServicePricipalName (id: $ServicePrincipalId) deleted."
+            $ServicePrincipalId = (
+                az ad sp list `
+                    --filter "displayname eq '$SubscriptionDeploymentServicePricipalName'" | ConvertFrom-Json
+            ).appId
+
+            if ($ServicePrincipalId) {
+                Write-Information "Deprovisioning $SubscriptionDeploymentServicePricipalName Service Principal (id: $ServicePrincipalId) ..."
+
+                az ad sp delete --id $ServicePrincipalId
+
+                Write-Information "Service Principal $SubscriptionDeploymentServicePricipalName (id: $ServicePrincipalId) deleted."
+            }
+        }
     }
 }
